# Performance improvements - Planner