--- a/go/vt/vtgate/planbuilder/operators/dml_with_input.go
+++ b/go/vt/vtgate/planbuilder/operators/dml_with_input.go
@@ -50,9 +50,6 @@ func (d *DMLWithInput) Inputs() []Operator {
 }
 
 func (d *DMLWithInput) SetInputs(inputs []Operator) {
-	if len(inputs) < 2 {
-		panic("unexpected number of inputs for DMLWithInput operator")
-	}
 	d.Source = inputs[0]
 	d.DML = inputs[1:]
 }